Reorganize style lib
