+++ src/main/lib/styled/index.ts
@@ -1,0 +1,4 @@
+import styled from './styled'
+
+export default styled
+export * from './themes'
