+++ src/main/lib/styled/themes/default.ts
